requirements.txt
discord.py[voice] @ git+https://github.com/Rapptz/discord.py.git@master#egg=discord.py
asyncpg
